requirements.txt
openai==1.30.1
openpyxl==3.1.2
pandas==2.2.2
pandas-stubs==2.2.2.240514
pysocks==1.7.1
python-dotenv==1.0.1
requests
pillow
requests_toolbelt